requirements.txt
flask==3.1.3
werkzeug==3.1.7
gunicorn==25.3.0
tensorflow==2.16.1
tf_keras==2.16.0
numpy==1.23.5
pillow==12.1.1
h5py==3.14.0
requests==2.33.0
whitenoise==6.7.0
itsdangerous==2.2.0
jinja2==3.1.6
click==8.3.1
protobuf==4.25.9
absl-py==2.4.0
flatbuffers==25.12.19
gast==0.4.0
google-pasta==0.2.0
grpcio==1.78.0
opt_einsum==3.4.0
packaging==26.0
six==1.17.0
termcolor==3.3.0
typing_extensions==4.15.0
wrapt==1.14.2
ml-dtypes==0.3.2